fix(promo): Queue a dummy job to execute previous jobs after a proccess reboot
Bullmq has a strage behaviour, probably a bug.

After a process reboot, the delayed jobs previously queued, are not processed by the QueueScheduler when the time comes.

Only after a new job is scheduled, the delayed ones, get executed. We intentionally queue this dummy job to keep on serving tasks.